backend/requirements.txt
crewai==1.11.1
crewai-tools==0.17.0
langchain==0.3.14
langchain-openai==0.3.1
langchain-community==0.3.14
langchain-chroma==0.6.3
chromadb==0.6.3
openai==1.59.9
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.20
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.10.5
pydantic-settings==2.7.1
tiktoken==0.8.0
sentence-transformers==3.3.1
faiss-cpu==1.9.0.post1
pypdf==5.1.0
python-dotenv==1.0.1
sqlalchemy==2.0.37
aiosqlite==0.20.0
httpx==0.28.1
pytest==8.3.4
langchain-huggingface==0.1.2
huggingface_hub==0.27.0
anthropic==0.40.0
